Refactors data fetching with Result type
Improves error handling and data consistency by introducing the `Result` type to the data fetching process.

This change replaces the previous `CriptoRepository` with `CoinsMarketsRepository` and `CoinsMarketsRepositoryRemote` to provide a more specific implementation for fetching coin market data. It also introduces `ApiClient` for managing API requests and data parsing, and updates the `DioHttpClient` to return `Result` for all HTTP methods.

The `CryptoModel` is renamed to `CoinsMarketsModel` to reflect its purpose more accurately.