Enables searching coins by name
Allows users to search for specific cryptocurrencies by name.

This change introduces a search functionality that filters the list of coins based on user input. It includes updating the API client and repositories to accept a 'names' parameter to filter the returned coin markets.

Also, adds l10n support for new strings used in the search feature.