Refactors data fetching and test implementation
Simplifies data fetching in coin details and home screens by
adjusting the parameters passed to the repository methods and
using the locale to define the currency.

Improves test reliability by using Key instead Icon to find the widget
and skip a non-stable test.

Fixes a bug in favorite list that always was empty.